Separate connection code and keep a static amount of worker threads
