+++ Cargo.toml
@@ -25,8 +25,6 @@ threadpool = "1.8.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 
-dotenv = "0.15.0"
-
 thiserror = "1.0.65"
 
 tokio = { version = "1.41.0", features = ["full"] }
