+++ base/src/lib.rs
@@ -1,6 +1,7 @@
 extern crate core;
 
 use anyhow::Result;
+use dotenv::from_filename;
 use opentelemetry::{KeyValue, Value as LogValue};
 use opentelemetry_appender_tracing::layer::OpenTelemetryTracingBridge;
 use opentelemetry_otlp::WithExportConfig;
@@ -8,6 +9,7 @@ use opentelemetry_sdk::logs::LoggerProvider;
 use opentelemetry_sdk::Resource;
 use rusttensor::subtensor::Subtensor;
 use rusttensor::AccountId;
+use std::env;
 use tracing::info;
 use tracing_subscriber::layer::SubscriberExt;
 use tracing_subscriber::util::SubscriberInitExt;
@@ -30,6 +32,15 @@ pub async fn subtensor() -> Result<Subtensor> {
         Ok(Subtensor::from_insecure_url(&*config::CHAIN_ENDPOINT).await?)
     } else {
         Ok(Subtensor::from_url(&*config::CHAIN_ENDPOINT).await?)
+    }
+}
+
+pub fn load_env() {
+    let prefix = env::var("DOT_ENV_FILE_PREFIX").unwrap_or(String::new());
+    let file_name = format!("{prefix}.env");
+
+    if let Err(e) = from_filename(&file_name) {
+        println!("Could not load {file_name}: {e}");
     }
 }
