+++ miner/src/serve_axon.rs
@@ -1,10 +1,8 @@
 use std::net::{IpAddr, SocketAddr};
 
 use clap::Parser;
-use dotenv::dotenv;
-use neuron::{config, subtensor};
+use neuron::{config, load_env, subtensor};
 use rusttensor::axon::{serve_axon_payload, AxonProtocol};
-use rusttensor::subtensor::Subtensor;
 use rusttensor::wallet::{hotkey_location, load_key_seed, signer_from_seed};
 
 #[derive(Parser)]
@@ -15,9 +13,7 @@ struct Cli {
 
 #[tokio::main]
 async fn main() {
-    if let Err(e) = dotenv() {
-        println!("Could not load .env: {e}");
-    }
+    load_env();
 
     let args: Cli = Cli::parse();
